+++ src/ConsolidateEmailInvoices/Configuration/EmailMappingSettings.cs
@@ -1,6 +1,0 @@
-namespace ExtractLoadInvoices.Configuration;
-
-public class EmailMappingSettings
-{
-    public Dictionary<string, string> SenderToFolderMap { get; set; } = new();
-}