+++ src/ConsolidateEmailInvoices/Services/IEmailProcessor.cs
@@ -1,8 +1,0 @@
-using ExtractLoadInvoices.Models;
-
-namespace ExtractLoadInvoices.Services;
-
-public interface IEmailProcessor
-{
-    Task<ProcessingResult> ProcessEmailsFromSenderAsync(string senderEmail, string destinationFolder);
-}